+++ cpp/opendnp3/src/opendnp3/app/APDUBuilders.h
@@ -30,15 +30,15 @@ namespace build
 {
 	// -------- requests -------------
 
-	void ClassRequest(APDURequest& request, FunctionCode code, int classMask, uint8_t seq);
+	void ClassRequest(APDURequest& request, FunctionCode code, uint8_t classMask, uint8_t seq);
 
-	void WriteClassHeaders(APDURequest& request, int classMask);	
+	void WriteClassHeaders(APDURequest& request, uint8_t classMask);
 
-	void ReadIntegrity(APDURequest& request, int classMask, uint8_t seq = 0);
+	void ReadIntegrity(APDURequest& request, uint8_t classMask, uint8_t seq = 0);
 
 	void DisableUnsolicited(APDURequest& request, uint8_t seq = 0);
 
-	void EnableUnsolicited(APDURequest& request, int classMask, uint8_t seq = 0);
+	void EnableUnsolicited(APDURequest& request, uint8_t classMask, uint8_t seq = 0);
 
 	void ClearRestartIIN(APDURequest& request, uint8_t seq = 0);
